chore: move entities types
