+++ src/entities/albums/index.ts
@@ -1,3 +1,3 @@
 export * from "./api/albumsApi"
-export * from "./api/albumsApi.types"
+export * from "./model/albumsApi.types"
 export { AlbumCard } from "./ui/AlbumCard"
