+++ src/entities/albums/model/albumsApi.types.ts
@@ -1,0 +1,5 @@
+export type Album = {
+  userId: number
+  id: number
+  title: string
+}
