+++ src/entities/photo/index.ts
@@ -1,1 +1,2 @@
 export { PhotoCard } from "./ui/PhotoCard"
+export * from "./model/photosApi.types"
