+++ src/entities/photo/model/photosApi.types.ts
@@ -1,9 +1,3 @@
-export type Album = {
-  userId: number
-  id: number
-  title: string
-}
-
 export type Photo = {
   albumId: number
   id: number
