+++ src/entities/post/index.ts
@@ -1,5 +1,5 @@
 export { PostCard } from "./ui/postCard/PostCard"
 export * from "./api/postsApi"
-export * from "./api/postsApi.types"
+export * from "./model/postsApi.types.ts"
 export * from "./model/postsSlice.ts"
 export * from "./model/postsSelectors"
