+++ src/entities/post/model/postsApi.types.ts
