+++ src/widgets/postList/index.ts
@@ -1,2 +1,2 @@
 export { PostList } from "./ui/PostList.tsx"
-export { type Post } from "../../entities/post/api/postsApi.types.ts"
+export { type Post } from "../../entities/post/model/postsApi.types.ts"
